Fix server REST api

+++ server.js
@@ -25,10 +25,33 @@ app.get('/api/bug', (req, res) => {
         })
 })
 
+// GET a bug
+app.get('/api/bug/:bugId', (req, res) => {
+    const bugId = req.params.bugId
+    bugService.get(bugId)
+        .then(bug => {
+            let bugsVisited = req.cookies.bugsVisited || []
+            bugsVisited.push(bug._id)
+            bugsVisited = Array.from(new Set(bugsVisited))
+            res.cookie('bugsVisited', bugsVisited, { maxAge: 60 * 1000 })
+            if (3 < bugsVisited.length) return res.status(401).send('Wait for a bit')
+            return res.send(bug)
+        })
+        .catch(err => {
+            console.error(err)
+            return res.status(400).send(`Cannot get bug with id=${bugId}`)
+        })
+})
+
 // PUT a bug
-app.put('/api/bug', (req, res) => {
+app.put('/api/bug/:bugId', (req, res) => {
+    const bugId = req.params.bugId
+    if (! bugId) {
+        res.status(400).send(`Cannot put bug with no id`)
+        return
+    }
     const updatedBug = {
-        _id: req.body._id,
+        _id: bugId,
         title: req.body.title,
         severity: req.body.severity,
         description: req.body.description,
@@ -39,7 +62,7 @@ app.put('/api/bug', (req, res) => {
         })
         .catch(err => {
             console.error(err)
-            res.status(400).send(`Cannot put bug id=${req.body._id}`)
+            res.status(400).send(`Cannot put bug with id=${req.body._id}`)
         })
 })
 
@@ -60,35 +83,20 @@ app.post('/api/bug', (req, res) => {
         })
 })
 
-// GET a bug
-app.get('/api/bug/:bugId', (req, res) => {
-    const bugId = req.params.bugId
-    bugService.get(bugId)
-        .then(bug => {
-            let bugsVisited = req.cookies.bugsVisited || []
-            bugsVisited.push(bug._id)
-            bugsVisited = Array.from(new Set(bugsVisited))
-            console.log(bugsVisited)
-            res.cookie('bugsVisited', bugsVisited, { maxAge: 60 * 1000 })
-            if (3 < bugsVisited.length) return res.status(401).send('Wait for a bit')
-            return res.send(bug)
-        })
-        .catch(err => {
-            console.error(err)
-            return res.status(400).send(`Cannot get bug id=${bugId}`)
-        })
-})
-
 // DELETE a bug
 app.delete('/api/bug/:bugId', (req, res) => {
     const bugId = req.params.bugId
+    if (! bugId) {
+        res.status(400).send(`Cannot put bug with no id`)
+        return
+    }
     bugService.remove(bugId)
         .then(bug => {
             return res.send(bug)
         })
         .catch(err => {
             console.error(err)
-            return res.status(400).send(`Cannot remove bug id=${bugId}`)
+            return res.status(400).send(`Cannot remove bug with id=${bugId}`)
         })
 })
